[Update] show all query data columns
