+++ frontend/src/utils/index.ts
@@ -1,5 +1,6 @@
-import { OperationMap, OperationStepMap } from '../types/operations';
+import { OperationDataMap, OperationMap, OperationStepMap } from '../types/operations';
 import { List } from 'immutable';
+import { ColumnList } from '../types/sources';
 
 export * from './api';
 export * from './localForage';
@@ -18,3 +19,20 @@ export const getSourceIDFromOperation = (operation: OperationMap): string | unde
 
 export const formatString = (name = '') =>
   name.split('_').map(word => `${word.charAt(0).toUpperCase()}${word.substr(1)}`).join(' ');
+
+export const getCustomColumns = (item: OperationDataMap, columns: ColumnList): string[] => {
+  const columnKeys = columns.map(column => column.get('name')).toJS() as string[];
+  const keysToIgnore = [ 'id', 'row_no' ];
+  const customKeys: string[] = [];
+  const iterator = item.keys();
+  let key = iterator.next();
+  while (!key.done) {
+    const { value } = key;
+    if (columnKeys.indexOf(value as string) === -1 && keysToIgnore.indexOf(value as string) === -1) {
+      customKeys.push(value as string);
+    }
+    key = iterator.next();
+  }
+
+  return customKeys;
+};
